refactor(website): extract projects domain into projects-db.ts (G-SIZE03)
Split website-db.ts (2890 lines) by extracting the Projects domain into a dedicated projects-db.ts module.
This reduces website-db.ts to 2218 lines, achieving G-SIZE03 goal (≤3000 lines).

Changes:
- website/src/lib/projects-db.ts: 803 new lines (Projects, SubProject, ProjectTask, etc.)
- website/src/lib/website-db.ts: reduced by 672 lines (2890 → 2218)
- 24 import statements updated to use projects-db directly
- Temporary re-export in website-db.ts for backward compatibility

The refactoring maintains full functional equivalence:
- No behavior changes, only code organization
- All Projects CRUD operations preserved
- Portal/customer access patterns intact
- Meeting relations co-located with Projects (logical grouping)

Quality Gates:
✅ G-SIZE03: 2218 lines (target ≤3000)
✅ TypeScript: no compilation errors

[T001359]
