+++ website/src/pages/api/admin/subprojekte/create.ts
@@ -1,6 +1,6 @@
 import type { APIRoute } from 'astro';
 import { getSession, isAdmin } from '../../../../lib/auth';
-import { createSubProject } from '../../../../lib/website-db';
+import { createSubProject } from '../../../../lib/projects-db';
 import { siteRedirect } from '../../../../lib/redirect';
 
 export const POST: APIRoute = async ({ request , locals }) => {
