+++ website/src/pages/api/portal/projekte.ts
@@ -1,6 +1,6 @@
 import type { APIRoute } from 'astro';
 import { getSession } from '../../../lib/auth';
-import { listProjectsForCustomer } from '../../../lib/website-db';
+import { listProjectsForCustomer } from '../../../lib/projects-db';
 
 export const GET: APIRoute = async ({ request }) => {
   const session = await getSession(request.headers.get('cookie'));
