+++ website/src/pages/api/portal/questionnaires/[id]/answer.ts
@@ -1,6 +1,6 @@
 import type { APIRoute } from 'astro';
 import { getSession } from '../../../../../lib/auth';
-import { getCustomerByEmail } from '../../../../../lib/website-db';
+import { getCustomerByEmail } from '../../../../../lib/projects-db';
 import { getQAssignment, upsertQAnswer, updateQAssignment } from '../../../../../lib/questionnaire-db';
 
 export const PUT: APIRoute = async ({ request, params }) => {
