+++ website/src/pages/api/portal/questionnaires/index.ts
@@ -1,6 +1,6 @@
 import type { APIRoute } from 'astro';
 import { getSession } from '../../../../lib/auth';
-import { getCustomerByEmail } from '../../../../lib/website-db';
+import { getCustomerByEmail } from '../../../../lib/projects-db';
 import { listQAssignmentsForCustomer } from '../../../../lib/questionnaire-db';
 
 export const GET: APIRoute = async ({ request }) => {
